Fix bazel dependencies for cmk.agent-receiver
This change fixes dependency issues during build in the same way
it was done for cmk.trace in 6a130a9cc2520a5257a0c030f8600c1864a598df.

It's a wild guess that this will fix our current CentOS 8 issues.
If this commit does not fix it, please revert. In case it actually
fixes the issue, this approach should be ported to the other cmk
packages.

Change-Id: Ic86f3a7b680e49f8f44709eb0f64db945b1821e2