17240 splunk: migrate health plugin to v2 architecture
In addition to the migration, I took the liberty to adapt the summary
lines to be more useful to the user. In the legacy plugin, all the
service feature health statuses would be grouped together into one
multiline string. Now in the migrated plugin, the features are
summarized alongside their respective service.

Change-Id: I4fff72dbb6b17f9da819b2b567e5a251c459b484