16121 FIX Fixed CMC crash when sending illegal metrics to InfluxDB
When the CMC tried to send to a metric to an InfluxDB where the metric name
was malformed (e.g. starting with an underscore), the CMC crashed.  This
has been fixed, such metrics are now simply forwarded to the InfluxDB.  The
latter will probably be a bit unhappy about such metric names, but this can
be figured out via its logs, and the corresponding check can be fixed.

Note that in a future Checkmk version, we will probably drop such metrics
silently, logging this only for higher log levels to avoid log spam at the
default log settings.

Change-Id: Iadbd96ae2c7d14acbb92103a1173d86f45d1ae8c