17246 FIX update_action: check for connector key in user configs before accessing it
We previously implemented an update action that would check for invalid
ldap connection ids in the user configs.  However, this change didn't
take into account that a user config doesn't always have a connection
id and would cause a crash on updating.
This werk fixes this issue by checking first if the user config has
a connection id before accessing it.

SUP-18845

Change-Id: I2a42ee99b4d21ebccf42e878ac68e57cbc34af62