Add internal onboarding endpoint
For the CSE we use a separate onboarding guide. The guide is loaded per
checkmk instance and shown on the first login. Here we want to allow us
to load the secret for an automation user.

The exact user has to be configured. This will be done during setup of
the site in the saas  platform. We are creating a user with minimal
permissions. Here for testing we use the automation user.

Change-Id: I7095a73229b76c2d800a5bc58a8b9f09cf90a54f
JIRA-Ref: SAASDEV-2253