17241 Support Azure PostgreSQL Flexible Server
Microsoft is retiring the Azure resource [1] "Database for PostgreSQL
single server". With this Werk, we now support monitoring the
recommended Azure resource "Database for PostgreSQL flexible server".

In the rule "Microsoft Azure" under "Azure services to monitor", users
can now select the new option "Database for PostgreSQL flexible server".
Note that the former option "Database for PostgreSQL" was renamed to
"Database for PostgreSQL single server" and stays in place.

The metrics monitored for flexible servers correspond to those monitored
for single servers and the same checks are used. See the check plugin
catalog [2] for more details.

[1]: https://learn.microsoft.com/en-us/azure/postgresql/single-server/whats-happening-to-postgresql-single-server
[2]: https://checkmk.com/integrations?distributions%5B%5D=check_mk&distributions%5B%5D=check_mk_cloud&search=azure_postgres

CMK-17959

Change-Id: I7bbdbb294dfcd3001b1604559ae37fbf7f773548