17255 Fix select acknowledgement comment removal
When selecting and removing Acknowledgement-type comments, all
non-persistent acknowledgement comments were removed for the respective
host or service, even if they were not selected.

With this werk, only the selected comments are removed in every case.
This allows you to selectively remove acknowledgement comments without
resetting the acknowledgement status of the host or service, as long as
other acknowledgement comments are present.

The acknowledgement status is reset once all acknowledgement comments
are removed.

SUP-17978

Change-Id: I65b1cfad50ae5608fb49268603dcf5fef9009d68