Change order of jaeger start script
Recently the liveproxyd stop procedure did not work correctly because
opentelemetry was blocking the correct shutdown of the process.

This change starts jaeger earlier and stops it later so that all other
services which may depend on jaeger can rely on it being available.

This should not be necessary, but is currently needed because of
shutdown and timeout related bugs in the opentelemetry exporter
implementation (see also CMK-18177).

Change-Id: Id6d990b931a7ebb9c5ec38e9dd17cfb2c324f2a2