testlib.agent: Do no call daemon.kill() after killing it via sudo
Apparently on some systems it may happen that the daemon is not
killed before daemon.kill() is called. If that happens, an error
message "Operation not permitted" is returned. This change removes
the daemon.kill() call, as it is pretty much pointless anyway,
since sudo privileges are required to kill a sudo process.

Change-Id: I3ae42e24f58a15ebe0fcca04053537c600b43023