15029 Improve setting up new least privileged user roles
Add new role "Role without any default permissions". The role is not indented to be used
directly. Instead it can be used to create new roles with only the
minium required permission. Even on updates you can be sure that the
derived role does not suddenly get new unexpected permissions.

Change-Id: I4f523edc6fe8eef8998920f761ec55fa506fd89a