Pipfile: update marshmallow and fix associating BaseSchema
The base.BaseSchema class is tightly coupled with its associated Marshmallow class and relies on private methods that were modified in the latest Marshmallow release (https://github.com/marshmallow-code/marshmallow/commit/7609530d0a61e6e5ba89769897e2b1908b881a50). This coupling has hindered the ability to update the Marshmallow package. This commit revises the class to align with the new implementation in the latest release, though it should be viewed as a workaround.

Change-Id: Icedb4e7bbdddef0d83af86e6d0bf527e58cb7378