17256 FIX Remove option to show help in lists
Previously, it was possible to display help boxes in ruleset lists like
the ones in the service monitoring rule selection page. The resulting
view had a number of visualization issues which made usage difficult.

With this werk, we remove the option entirely in an attempt to slim
down the global settings page.

Change-Id: I45ebe1c30681e73526d5a9723a529a1571efbd9d