17235 FIX cpu_utilization: allow total CPU utilization to be set above 101%
Before this werk, the "High utilization at" level option within the
"Levels over an extended time period on total CPU utilization" target
was limited to a maximum of 101%. However, in environments like containers,
the total CPU utilization can exceed this threshold. This werk therefore removes
the upper limit for the total value.

SUP-19611

Change-Id: I744f002a4d3f2e04ef372fabf41dea5485938382