vue: Fix component props typing when re-exporting
We encountered a bug with vue/ts in our configuration:
If we re-export a vue component, it will be typed to any
if used inside Vue's <template>.
Disabling this typescript rule restores the expected typing.

Github issue link:
https://github.com/vuejs/core/issues/11732

This also uncovered that radix didn't adhere to our typing rules,
we silenced those errors.

CMK-18584

Change-Id: Ieeef6f3d7cadcbfbc0366ee67972e8b7dd9e16c9