16352 Support Diagnostics: Add rotated log files to the dump
With this werk, it's possible to select rotated logfiles for the dump, for example update.log and update.log.1
In the "Pack only xyz sensitivity files" and "Pack all files" modes, both are are added automatically.
In the "Select individual files from the list" mode, the user can select them separately.

Background: In rare cases, the .log files were empty, because the logging process still wrote into the .log.1 file.
Change-Id: Id2f02b36eabe1052615876ce8b8c18ca844f6d71