17355 raritan_pdu_plugs.py: respect user defined parameters
Currently, the check disregards the custom parameters set by the user,
always returning the state defined by the SNMP mapping. With this
change, the service now compares the current state with the required
state defined by the user in the check parameters. If there is a
mismatch, a "CRIT" result will be yielded, comparing the actual and
expected plug state. Otherwise, an "OK" result will be yielded with the
plug state.

If these parameters are not explicitly set for a service, the service
will compare the current state with the state set during service
discovery.

CMK-18917

Closes: #753
Co-authored-by: Logan Connolly <logan.connolly@checkmk.com>
Change-Id: I778f8d4aaf73ee9dad5e684ade88bf7141142b5f