16998 FIX SLA alert levels: show CRIT state if SLA alerts levels are CRIT but SLA not broken yet
When configuring SLA alerts levels only the WARN state was shown in a view,
in case the alert levels for warning have been reached.

The CRIT alerts levels were shown as OK as long as the SLA was not broken.

Now this already shows up as CRIT, even though the SLA is not yet broken.

SUP-19972

Change-Id: Ie97ddbe410a468c946efbd9db398bb637a4e154e