Refactoring: Simplify the handling of expected regular messages a bit
Apart from simplifying a message signature, we prepare the code for a change
of the "expect" key of a rule: Instead of being an optional key of type
"Expect", it shoud better be a mandatory key of type "Expect | None".  This
will ease a later change from TypedDicts to dataclasses.

Change-Id: Id5c169163d0bb090064826628f31950676ff16ae