Move cmk.utils.daemon to cmk.ccc.daemon
The module is a collection of functions to help daemonizing a process
which is used in various of places.

Moving this module out of cmk.utils is needed to free cmk.cee.mknotifyd
from the dependency to cmk.utils so that we can move cmk.cee.mknotifyd
to a separate package.

Change-Id: I074f8909d04acbb01e96c33ce7b07e045b0c88ad