Move cmk.utils.profile to cmk.ccc.profile
The module provides a small helper to profile python code. It is used
by various parts of Checkmk where profiling can be enabled ad-hoc.

Moving this module out of cmk.utils is needed to free cmk.cee.mknotifyd
from the dependency to cmk.utils so that we can move cmk.cee.mknotifyd
to a separate package.

Change-Id: I86e7949f488ad9822a13eae5f78b275e77924570