16566 FIX Standardize notification spooler log level configuration
The notification spoolers log level was configured differently than other
services. Previously users had the choice between:

* Normal logging (only startup, shutdown and errors)
* Verbose logging (also spooled notifications)
* Debugging (log every single action)

This was changed to the standard log levels which we also use in all other services.

The command line flag `-v` of the `mknotifyd`, which could be used to override
the configured log level was replaced with the `--log-level=LEVEL` argument.
While the change to this command line flag is an incompatible change, we don't
rate this as a change that is worth marking the werk to be incompatible. In the
end it's only an option used for manual debugging.

With this change the notification spooler is using the standard python
logging configuration. There is nothing Checkmk specific about it anymore.

Change-Id: I1f5dda58772fece288f0a7ae72c171358e98edc6