17267 REST-API: Introduce include_links and include_extensions to list endpoints
This Werk introduces two flags to most list endpoints.

The first flag `include_links` was already present for list host configs. It
controls whether the links of the individual values should be included.
This flag is enabled by default.

The second flag `include_extensions` toggles the inclusion of the extensions,
which contain most of the attributes. This flag is also enabled by default.

Both of these flags give users options to improve performance and reduce
response sizes.

Change-Id: Ia7f7b6798aba7020dbbbe0e49f0b48e127527d5a
JIRA-Ref: CMK-18835