16870 FIX vsphere: Determine snapshot age based on system time
Previously, snapshot age was calculated based on the Checkmk server system time.
This led to incorrect values when the time zone of the vSphere system was different
than the time zone of the Checkmk server.

Now, the age calculation is based on a vSphere system time.

Note: This became an issue in 2.2 with werk 17197.

SUP-21356

Change-Id: I1129cbae32c114da8f29ce7c9855482417964164