Make cmk.utils.event_context independent
We are about to move several event processing related modules to a
dedicated package. As last preparational step, we need to get rid of the
cmk.utils.hostaddress.HostName dependency.

The idea of this change, which will get relevant for other packages as
well, is:

1. The package defines its own type for HostName.
2. At the boundaries of the package we need to convert between the
   package's HostName type and the other host name types (like
   cmk.utils.hostaddress.HostName).

In the package specific type, we assume that the value was already
parsed, so we can simply use a NewType.

Change-Id: I1efaa94fbc0201f07a2f6ea0131cbe6172fa9570