Fix filtering of package releated python files in gerrit job
The filtering was broken since beginning of the year (since commit
ba8b5eda71ffff1cef35c8a9950fca2f5a53e08f) which made the jobs care for
package related files which they should not.

For example, the global "Python Format" stage cared for files in the
`cmk-trace` package, which is the responsibility of the
"Package cmk-trace" stage.

Change-Id: I0a070ea55e829b03f7edee77294c222dbf7a2b35