17035 mysql_replica_slave: Adapt mk_mysql for MySQL version 8.0.22 and above
From MySQL 8.0.22, SHOW SLAVE STATUS is deprecated and the alias SHOW REPLICA STATUS should be used instead.
The statement works in the same way as before, only the terminology used for the statement and its output has changed.

For the sake of compatibility, the service name will stay the same 'MySQL Slave'.
There is no user intervention required.

SUP-20165

Change-Id: I87a1ce806976475a0eebb76309ac4ae509bd7642