move from composition api to options api
otherwise the scoped styling does not work as expected.

currently the `data-v-` attributes will not be added as the style is
empty. they are only added if really needed.

also removed the model for data, as the component should not change the
data and thus does not need two-way-data-bindings

CMK-19021

Change-Id: I3bcff5b806522839f38ce9471bdb975d697fa634