Reduce time_machine usage a tiny bit
Don't conjure up the current time deep down the call stack and use
time_machine in tests later to fix that API deficiency.  Instead, pass down
the current time from the top, i.e.  check_logwatch_ec_common(), and use
that aftwards.

Note: With a slightly saner organization of the code into classes and
methods, the threading of the timestamp could be reduced quite a bit, but
that's something for another change.  ;-)

Change-Id: Id6931f15b85ebf5c4638e96255f20184800792aa