tests/conftest: Enrich TimeoutError exceptions with top output
To simplify the debugging of timeouts in the CI, this change adds resource usage
information to TimeoutError exceptions raised by the tests.

This is done via the pytest_exception_interact hook. We therefore always define
that hook now and add the necessary information to the exception message for
the relevant exception types (TimeoutError, PWTimeoutError).

CMK-18910

Change-Id: I7f6b346610fbceb09343ac1067f1a233e8d06822