16123 FIX Use the original message text for rewriting the comment field
Werk 16534 made the rewriting of the comment field use the comment field
itself as the basis. For this field, it doesn't really make sense, because
rewriting is the only way to fill that field, so e.g. \0 was always empty.

With this change, we revert to the old behaviour where all match groups
refer to the original message text.

Change-Id: I761db0445b485da62c30656b42bcc65ed30d1151