SSC backend: remove 'get_active_service_description'
This function originated from a time when it was cheaper
to only compute the description, but this is not the case
anymore. These two functions only differ in the check that
the description is not empty and the deduplication.

In both callsites we actually want that, though (or it
makes no difference).

Change-Id: Ib7baf5db49d2320258d6a7711aa13c14aea9652e