PRO TIP: migrate check_levels from v1 to v2
The parameters that are accepted by the v2.check_levels
are exactly as created by the new [Simple]Levels form spec.

In order to start using the new API's check_levels, the
ruleset needs to be migrated to the new API.
The heavy lifting is done by the migration function
passed in line 41 of the new file. It will convert all
existing configurations upon update.

There are four such functions exposed by the rulesets API:

 * migrate_to_float_simple_levels
 * migrate_to_float_levels
 * migrate_to_integer_simple_levels
 * migrate_to_integer_levels

In our case we are dealing with a simple tuple of floats
(and no predictive levels), so we chose
`migrate_to_float_simple_levels`.

More on the migration of rulesets can be found in
commit 4b3bd78f013a86cf8da710ec42ef1fc7eee06ba8.

In the check plugin, we only replaced the `check_levels`
function, and slightly adopted the default parameters.

More on the allowed values for the levels can be found
in the Sphinx documentation on the agent based API and
the ruleset API. The Sphinx documantation can be found
in a running site, in the "Plugin API reference" section
of the "Help" menu.

Change-Id: I8f529a9ad5bc5a0145489cc22436850f9f548b40