Sphinx autodoc: Clean up warnings and enable fail-on-warning
Autodoc creates some warnings like this, which are not helpful and
silently produce empty API doc pages. To fix that, we ignore these
missing dependencies and suppress the other warnings which are shown
after the first step.

Now that we are warning clean, we enable `--fail-on-warning` to catch
any new warnings that are introduced.

Change-Id: Icaff604b9d3bae267f1e1b02ed1ad44f9b4415fc