16895 FIX omd: curl wrapper not used for all supported versions of SLES and EL
Some OMD makefile rules have to be (de-)activated for certain versions of SLES or EL
(Enterprise Linux), but would not respect all affected versions due to the redundant
implementation of the check. This led to {{curl}} not being wrapped as needed for SLES15-SP3.

This change makes use of string prefixes rather than explicit listing of all supported
versions of e.g. SLES15[-sp*], which fixes the mentioned issue and also takes care of
future releases of SLES15 or EL based distros.

SUP-20320

Change-Id: Ifc7eced25e02ee2be34ba27de76671aa5e17a551