ship css for cmk-frontend-vue
this is a bigger change and does multiple things at once, but the most
important thing is that previously we did not ship the css styles for
cmk-frontend-vue. they were visible with the `vite dev` server, but the
css was not included in the builds.

this is the first time we make use of the manifest.json provided by the
vite build. this way we can copy the whole dist folder as is and
facilitate the cache busting mechanism provided by vite. the filenames
contain hashes, so we don't have to add a query to the urls.

biggest problem with this change was the relative external url of the
icons: it is resolved relative to the css definition. but this differs
between production build and development, as in production it is
compiled into a file and during development its inlined into the dom.
to solve this, the files are not directly accessed and no longer
external, so vite can find them on the disk. this was previously not
possible because of the theming mechanism in place, but the variables
resolved this problem, if you inline the apache redirection that
resolves icons not found in modern-dark with facelift.

the second problem was: how to identify the css file and where to put
it. in this change we use a new approach to utilize the manifest.json to
read the filenames from there. this way we can simply copy the whole
dist folder into the production build and read the filenames from there.
maybe we have to introduce some performance optimizations, as we now
read the manifest.json for each request.

also modify the logic when to add vue javascript files slightly:
previously we explicitly added vue as a js dependency, now we add the
vue javascript files whenever the main javascript is added.

CMK-19056

Change-Id: I168af7449db06ee4e679a6f8ab502277a398f8fa