17244 gerrit: monitor version of deployed instance
This check compares a currently deployed Gerrit instance with newer available releases.

Since Gerrit uses a release cycle similar to semantic versioning (https://semver.org), this check
can monitor when a major, minor, or patch release is available. With the service configuration, a
user can specify the specific type of release(s) they would like to be notified about.

For example, an administrator may only want to be notified when there is a new patch release, but
isn't concerned about major or minor releases.

CMK-17966

Change-Id: Ie57a515cca31fd06a46780eb5126a0e6216ff766