17258 Fix host not renaming when node of cluster
Due to a bug in our cache invalidation mechanism hosts were not renamed
when they were a) in a subfolder and b) a node of a cluster.

This commit fixes the issue by moving the Host.all() call to after the
hosts were renamed. This is only a workaround/hotfix and doesn't
address the underlying bug as this might require a bigger refactoring:

The folder._hosts is not cleared by the cache invalidation but even if,
it would not be invalidated everywhere as not every folder object is in
the subfolder hierarchy of the root folder.

SUP-18050

Change-Id: I14fd54a80c49466f034b4e83bf77646cec01c75e