17310 licensing: Changes in license compatibility
Previously, a Checkmk MSP site was usable with a Checkmk Cloud license, but not the other way around.
Since the Checkmk Cloud is now included in the Checkmk MSP, this changes to the following compatibility:

* Usable with a Checkmk Enterprise license: Checkmk Raw, Checkmk Enterprise
* Usable with a Checkmk Cloud license: Checkmk Raw, Checkmk Enterprise, Checkmk Cloud
* Usable with a Checkmk MSP license: Checkmk Raw, Checkmk Enterprise, Checkmk Cloud, Checkmk MSP

Note that this only refers to the compatibility when applying a license to a site, not whether sites of different editions can exist together in a distributed setup/monitoring.

If you plan to upgrade from Checkmk Enterprise to Checkmk Cloud or Enterprise Checkmk Cloud to Checkmk MSP, please upgrade your subscription first.

CMK-17181

Change-Id: I0d4b8c1164933eec60adaff7c957e74c0231c14f