Clean up edition specific PainterDowntimeRecurring
The previous implementation registered this painter in all editions.
There are built-in views shipped which contain this printer for all
editions.

For the Raw edition there was an edition specific renderer registered
which statically printed "(not supported)". Instead of registering a
printer displaying this static text, we now simply do not register the
painter at all, which removes the need for a edition specific monkey
patch. The view will then silently not display the printer, which should
be fine from the user's perspective.

Change-Id: Ic2b3db45d75208efcdc82b9d2cfad25578cb8e95