discovery: reorganize cluster service merging
We used to compute the transitions (new, vanished, ...)on the
individual nodes before computing the cluster aggregation.
This is unnecessarily confusing.
We now compute two aggregations (before/after) for the cluster
and then derive the transitions from these.

Also we now have the merged discovered labels of clusters
considered by the ruleset matcher.

This fixes CMK-18804 and CMK-17762 and unblocks SUP-6185.

Change-Id: Ib6ea13a2ea2ddd0b7a275fd8a28b859ae3e6542f