17259 Fix mismatching network interface and switch port services
When configuring network interfaces to be discovered by alias or
description (rule "Network interface and switch port discovery"), there
was a chance of a mismatch between the discovered service item and the
interface checked by the service.

This happened if the alias or description of one interface matched the
index of another interface. For example, if the interface discovery is
configured to discover interfaces by alias, then the service "Interface
5" is supposed to check the interface with the alias "5". Instead, it
checked the interface with the index 5.

SUP-20568

Change-Id: I030c4beffeb15465c77b73bfa74e90ba9c6fda3c