17037 FIX Adapt change from werk #17035 for MariaDB
With the change introduced with werk #17035, if a system was using a MariaDB version higher than 8, it would cause the MySQL DB service to go to unknown.
This was due to the fact that the command 'SHOW REPLICA STATUS' is not available in MariaDB version lower than 10.5.
The user needs to update the agent plugin "mk_mysql" in order for this to change to work as intended.

Change-Id: I4dc1e830c729cced06b77069e5291a13e14e1a3c