17330 Fix comment selection being site independent
When selecting comments in a view with comments from multiple sites,
rows were selected by comment id regardless of which site they were
from. This led to cases where trying to delete a single comment with
comment id `1` from one site led to all comments with the comment id `1`
being deleted from all sites.

With this werk, comment selection in view is site and id dependent,
allowing the user to delete single comments even if they share an id
with a comment on a different site.

SUP-20321

Change-Id: If76f87012f20f51d05397404e2d1ef7212c3c86d