Change audit log to be newline separated
In our SaaS environment we want to be able to read that file with other
tools but they rarely support the `\0` as delimiter. Therefore we're
changing it to `\n`.
We are still logging python objects here so parsing with JSON will not
work, also the audit log is unstable and might be subject to further
change. So do not rely on structure or content!

CMK-19258

Change-Id: Ie5a7756e8e4def6f07748025af3a17e3ce97c16d