17311 FIX mk_logwatch: Apply context to informational (OK) messages
With the `maxcontextlines` option the output of the mk_logwatch agent plugin can be reduced to a certain number of lines before and after relevant messages.
Previously the messages were seen as relevant if they had the level `W`(warning) or `C`(critical).
With this werk messages that are marked with `O` (OK/informational) are also considered relevant and will have the context applied.

SUP-20390

Change-Id: I2acf75ef9b92221243d2a36d8a0db82852320142