15518 Enable symmetric agent encryption on Solaris
This werk adds symmetric agent encryption on Solaris hosts. The
functionality is identical to the symmetric encryption for Linux hosts.

To use the current and save encryption scheme, OpenSSL >= 1.0.0, better
OpenSSL >= 1.1.1, must be available on the host.

Older encryption schemes can still be decrypted by the Checkmk site as for
Linux.

Change-Id: I92c0427a2a62b48142d7fc293cbfdf0cdd0e68ab