Enable tracing only if a tracing context is provided
Would be nice if we could enable tracing in all cases. But the current timeout and retry
logic of the Opentelemetry SDK does not work as intended, which can lead to longer hanging
executions in case the configured collector (e.g. Jaeger) is not reachable.

For now only enable tracing in case the caller provides a tracing context. This way we produce
traces only when the caller is interested in them.

This condition can be removed once the timeout and retry logic was fixed
(https://github.com/open-telemetry/opentelemetry-python/issues/4043).

Change-Id: Id9ce70cf3232a5e1bbe9497e8c30a568ac693b78