Silence pip-compile
Without this, we are totally drowning in a vast sea of useless output when
generating/testing the requirements files: For e.g. the toplevel we
currently get 4287 lines (roughly 72 DIN A4 pages) of output, for the tiny
cmk-agent-based we still get more than 2 pages.  This makes it basically
impossible to see actual problems, and the output is contained in the
generated requirements file, anyway.  Note that for a failing test we would
still see the diff, even with --quiet.

Change-Id: Ib2cd3e2f2f06e6bec066d3748924e5e19e503ffa