Introduce semgrep rule to warn about use of TypeAdapter
We had several performance regressions in the past because of the use of
TypeAdapter in often called code paths. We now add this rule to warn
developers about the use of TypeAdapter. After careful review, checking
the performance aspect, the developer can add an annotation next to the
TypeAdapter call to suppress the warning.

Besides the nature of the rule, this is also introducing the first semgrep
rule to the code base. The idea is to replace some of our custom pylint
checkers with semgrep rules in the near future. We will also integrate
semgrep with our CI in the next steps.

CMK-19526

Change-Id: Id464a5de0e06f13cd3141c5c671ea49cfae0b27b