fix: builtin hooks just forward exceptions
For builtin hooks we do not want to change the exception handling
behavior of the calling code.

The previous behavior of wrapping exceptions can result in surprising
behavior, for example the sidebar is using crash-reports to allow
reporting of exceptions. But an exception in a hook was not handled by
the crash reports. If we raise the plain exception however the sidebar
correctly generates a crashreport.

fixes: SAASDEV-2492
Change-Id: I7069877b14a68b0ebd89846cf74a9533b08a3373