17286 apc_ats_output: Add metrics and rulesets for current and power
Previously the monitoring of the current and power of an ATS only included whether their values were available and displaying them in the service summary.
Now they are emitted as metrics and thresholds configurable in the rule "APC Automatic Transfer Switch Output" are applied.

Additionally, load values of "-1", meaning that the ATS does not support this output, are now ignored.

Thanks to dnlldl for the contribution!
closes:#630
CMK-17700

Change-Id: I165305bd8a01e6f6a26fda1ddf09281ec672fa27