Replace code quality checks for left over print debug statement with semgrep rules
The semgrep checker has a better detection logic than our previous
pattern matching approach, so we need to clean up a few things in order
to land this checker.

Also extend the checker scope to the package directory to encurage the
clean up while moving the code to the new package hierarchy.

It still needs to be decided whether we want to keep the semgrep checks
on a repo level or move them down to the individual packages.

Change-Id: I1824a8dcf084cba9335f944233e4e5198117cd1c