test: evaluating driverstatus
It's not really clear, that running eval on the output of
```
tw_cli /"$TWAC" show drivestatus
```
is safe. In order to replace "{!DNAME}" with a POSIX compatible
version we would need to call `eval` a second time. $DNAME is output,
which comes from the device, and thus not really trusted. Probably ok
anyway, but let's be safe and use a different approach.

This commit only aims to test the functionality, which will refactored.

SUP-14161

Change-Id: Ib5aab185bf7e85376a7c4005d119bd9201477ad4