don't evaluate driverstatus
It's not really clear, that running eval on the output of
```
tw_cli /"$TWAC" show drivestatus
```
is safe. In order to replace "{!DNAME}" with a POSIX compatible
version we would need to call `eval` a second time. $DNAME is output,
which comes from the device, and thus not really trusted. Probably ok
anyway, but let's be safe and use a different approach.

SUP-14161

Change-Id: I7be9486b7a5632398c889eef004c8caedf142bfa