Clean up CME specific livestatus connection handling
So far we mostly use specific registries for dispatching edition
specific functionality. This feels acceptable for many parts, but not for
all. For example, it does not seem like a good approach to have a
registry just for a default livestatus connection object and one for the
CME livestatus connection object.

For this clean up, the commit introduces a Features object which is
attached to our WSGI app depending on the desired edition. To do so,
each edition creates it's Features object during registration. The
objects are then dispatched by our WSGI app during initialization.

This is also a test baloon to see if this approach is more maintainable
and can actually be a path towards clean edition aware unit tests. Let's
see how far we can get with this.

Change-Id: I36674782268cf77c3975ffb4fb3b21fd8289abe2