16568 FIX Fix CME specific validations on host and folder actions
CME specific validation to enforce the consistency of customer in a folder
hierarchy were not executed properly during the following actions since 2.3.0:

* Create host
* Create subfolder
* Edit folder
* Move host
* Move folder

Those validations ensure that a lower level of the folder hierarchy (host or
folder) can not have a site assigned which is owned by a different customer than
the levels above.

Due to the missing validation in previous 2.3 releases, it might have happened
that unintended configuration changes were made to the configuration. Such
possible inconsistencies need to be cleaned up manually.

Change-Id: I28c5ca82f552bef828d37934d174f2bd9b8a56a0