Make the control flow a tiny bit clearer
The "switch" is the last statement within the "while" loop, so doing a
"break" within the switch is equivalent to doing a "continue" for the loop.

The plan is to move some code into separate functions to split up the
monster functions, but to do so it's important to keep the control flow more
local.

Change-Id: I9f19f6bc5ca540e6749c4dff0d19b7d7269bc2b5