Trying hot fix for composition tests
Our current understanding is that we hand over the complex Folder
objects to the host rename background jobs. If such object holds a
reference to a lambda object, which the Folder object does since the
commit 498656c1462cc1c0fa7c3c3df7ac962a3ff89873, the pickle operation,
which is used to serialize the object for the background job, will fail.

In general it seems to be a good idea to remove references to such
complex objects in background job calls. Not only for maintainability,
but also for performance reasons. We'll try to move towards this
direction with a proper follow-up fix.

Change-Id: Ie7bc21fc52d3782d47d701907a7c3c671b1bf478