Simplify renaming background job start
Handing over complex objects to background jobs is not a good idea. Not
only in terms of maintainability, but also for performance reasons. The
objects and pickled between the processes, which has some traps in the
detail (e.g. lambda objects) and also has a performance impact.

In this specific case the problem occured, because the Folder object
holds lambdas since the commit 498656c1462cc1c0fa7c3c3df7ac962a3ff89873.

Change-Id: I302a2ac7ed05cc9d40ed22b972b1fc189da678d9