Fix host collection type: The automation call needs a list
Unfortunately the typing is not complete here so that mypy can not
uncover the issue. The SerializedSettings break the chain. A cleanup
would be a bit too much now.

Reason for back-picking this change: In a distributed setup with a 2.3 central
and 2.4 remote, the automation helper of the remote site complains when being
called from the central site:
```
Got invalid data:
Internal automation error: Error running automation call restart: 1 validation error for _AutomationPayload
args
  Input should be an instance of Sequence [type=is_instance_of, input_value={'test'}, input_type=set]
```

Back-picking this change should be risk-free, since already in 2.3,
`check_mk_local_automation_serialized` converts internally to a `list`:
```
new_args = list(args)
```

Change-Id: I6bf959fe86472c803dea36b81a797371814e80b2