Moved processLogEntries to where it belongs
The LogCache/Logfile "API" is weird and is used very differently in the
implementations of the "statehist" and "log" tables.  So let's move stuff
out of that "API" into its use sites and try to come up with a sane API
later. "No" API is better than a broken API. :-}

Change-Id: Id1436c2f19a456d651fe19332c7618e5e00311fb